Refactor CodeService class into AccessCodeService
