+++ app/core/services/__init__.py
@@ -1,3 +1,3 @@
-from .code import CodeService
+from .access_code import AccessCodeService
 from .jwt import JWTService
 from .session import SessionService
